Formik started collecting data properly and temporarily alerting them
