+++ frontend/matura/src/main.jsx
@@ -1,25 +1,27 @@
 import React from 'react'
 import ReactDOM from 'react-dom/client'
 
-import { createBrowserRouter, Navigate, RouterProvider } from "react-router-dom";
-import Login from "./Login.jsx";
-import { ChakraProvider } from "@chakra-ui/react";
+import {createBrowserRouter, Navigate, RouterProvider} from "react-router-dom";
+import LoginForm from "./LoginForm.jsx";
+import {ChakraProvider} from "@chakra-ui/react";
 
 const router = createBrowserRouter([
     {
         path: "/",
-        element: <Navigate to={'/login'}></Navigate>
+        element: <Navigate to={'/login'}/>
     },
     {
         path: "/login",
-        element: <Login/>,
+        element: <LoginForm/>,
     },
 ]);
 
-ReactDOM.createRoot(document.getElementById('root')).render(
-    <React.StrictMode>
-        <ChakraProvider>
-            <RouterProvider router={router}/>
-        </ChakraProvider>
-    </React.StrictMode>,
-)
+ReactDOM
+    .createRoot(document.getElementById('root'))
+    .render(
+        <React.StrictMode>
+            <ChakraProvider>
+                <RouterProvider router={router}/>
+            </ChakraProvider>
+        </React.StrictMode>,
+    )
